# contracts/openapi.v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Gorodki.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/auth/google": {
      "post": {
        "tags": [
          "Вход"
        ],
        "summary": "Вход через Google; новый игрок — по инвайту, 16+ и согласию",
        "operationId": "signInWithGoogle",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleSignInRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": [
          "Вход"
        ],
        "summary": "Новая пара токенов взамен refresh-токена",
        "operationId": "refreshSession",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": [
          "Вход"
        ],
        "summary": "Выход: отзывает все токены этого входа",
        "operationId": "logout",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GoogleSignInRequest": {
        "required": [
          "idToken",
          "ageConfirmed"
        ],
        "type": "object",
        "properties": {
          "idToken": {
            "type": "string"
          },
          "inviteCode": {
            "type": [
              "null",
              "string"
            ]
          },
          "ageConfirmed": {
            "type": "boolean"
          },
          "consentVersion": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          }
        }
      },
...
      "RefreshRequest": {
        "required": [
          "refreshToken"
        ],
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        }
      },
...
      "SessionResponse": {
        "required": [
          "accessToken",
          "refreshToken",
          "expiresIn",
          "isNewUser"
        ],
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "expiresIn": {
            "type": "integer",
            "format": "int32"
          },
          "isNewUser": {
            "type": "boolean"
          }
        }
      },
...
    }
  },
...
}